Update on work done
So a lot of things happened, but the project is by no means in a ready
state.
For one I moved on from Pistache and nlohmann json to cpprestskd, as
that seemed like it is much better supported and is crossplatform,
and above all else is very well documented.
However it's still very hard to get it to, start building(find all the
packages is cmake), compile, and run. I yet been able to get it to run.
Sooooooooooo that's that.